Pariing done tests, avoiding 32 machines for now.

+++ test/runtests.jl
@@ -8,7 +8,7 @@ end
 
 @testset "Boolean (Formulas)" begin
     ## Set up
-	init_logic(Int64(3))
+	init_logic(3)
     f1 = "x1 + x2 * x3"
     f2 = "(x1 * x3) ⊕ (x2 * x3)"
